feat: use real filler tile data for nodata regions
This provides consistent unbroken filling of colour across the map. It
relies on the fact that all the empty regions are the sea where we can
assign an average visible surface area. This average can then be used by
the shaders to dynamically adjust the heatmap colours to the surrounding
area. Previously we just hardcoded a static CSS hex value that remained
a gret colour no matter the zoom level or surroundings.