Digital Ocean provider for Atlas
This can run the entire world on an arbitrary number of Digital Ocean
machines.

It suffers one problem: the input `.bt` tiles have to be SSH'd up to
the VMs, which on my current Wifi is a serious bottle neck. Next commit
will be to pre-stitch the entire world's `.bt` tiles into an S3 bucket,
that way the worker machines only need to download the tiles over their
no doubt beefy ethernet connections.